pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "helix-context"
version = "0.3.0b5"
description = "Genome-based context compression for local LLMs"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [
    { name = "Michael Bachaud", email = "swiftwing21@users.noreply.github.com" },
]
keywords = ["llm", "context", "compression", "ollama", "genome", "rag"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "httpx>=0.27",
    "pydantic>=2.6",
]

[project.urls]
Homepage = "https://github.com/SwiftWing21/helix-context"
Repository = "https://github.com/SwiftWing21/helix-context"
Issues = "https://github.com/SwiftWing21/helix-context/issues"

[project.optional-dependencies]
accel = ["orjson>=3.9"]
embeddings = ["numpy", "sentence-transformers"]
cpu = ["spacy>=3.7"]
launcher = ["jinja2>=3.1", "psutil>=5.9"]
launcher-native = ["jinja2>=3.1", "psutil>=5.9", "pywebview>=5.0"]
ast = [
    "tree-sitter>=0.23",
    "tree-sitter-python>=0.23",
    "tree-sitter-rust>=0.23",
    "tree-sitter-javascript>=0.23",
    "tree-sitter-typescript>=0.23",
]
scorerift = ["scorerift"]
# CPU-resident semantic compression for gene content. Headroom by Tejas Chopra
# (https://github.com/chopratejas/headroom, Apache-2.0). See NOTICE.
codec = ["headroom-ai[proxy,code]>=0.5.21"]
dev = ["pytest", "pytest-asyncio"]
all = [
    "orjson>=3.9",
    "numpy",
    "sentence-transformers",
    "spacy>=3.7",
    "tree-sitter>=0.23",
    "tree-sitter-python>=0.23",
    "tree-sitter-rust>=0.23",
    "tree-sitter-javascript>=0.23",
    "tree-sitter-typescript>=0.23",
    "headroom-ai[proxy,code]>=0.5.21",
]

[project.scripts]
helix = "helix_context.server:main"
helix-launcher = "helix_context.launcher.app:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["live: requires Ollama running with at least one model"]
